pyproject.toml
[project]
name = "forexfactory-scraper"
version = "0.1.0"
description = "Gets calendar events from ForexFactory.com calendar"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "4x" }
]

dependencies = [
	"selenium",
	"nodriver",
	"python-dateutil",
	"tzdata",
	"pandas",
	"rich>=14.1.0",
	"forex-common",
]

[project.urls]
Repository = "https://github.com/4x/forexfactory-scraper"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
forex-common = { path = "../forex_common", editable = true }
